remove controller
