--- lib/app/shared/widget/account_select_box/controller/account_select_box_controller.dart
@@ -1,11 +1,0 @@
-import 'package:altme/wallet/wallet.dart';
-import 'package:bloc/bloc.dart';
-
-class AccountSelectBoxController extends Cubit<CryptoAccountData?> {
-  AccountSelectBoxController({CryptoAccountData? selectedAccount})
-      : super(selectedAccount);
-
-  void setSelectedAccount({required CryptoAccountData selectedAccount}) {
-    emit(selectedAccount);
-  }
-}
