+++ lib/app/shared/widget/widget.dart
@@ -1,4 +1,3 @@
-export 'account_select_box/account_select_box.dart';
 export 'add_account/add_account.dart';
 export 'altme_logo.dart';
 export 'app_version.dart';
