+++ lib/dashboard/home/tab_bar/tokens/token_page/widgets/token_list.dart
@@ -1,6 +1,5 @@
 import 'package:altme/app/app.dart';
 import 'package:altme/dashboard/dashboard.dart';
-import 'package:altme/dashboard/home/tab_bar/tokens/view/widgets/widgets.dart';
 import 'package:flutter/material.dart';
 
 typedef OnScrollEnded = Future<void> Function();
