+++ lib/withdrawal_tokens/view/insert_withdrawal_amount_page.dart
@@ -1,6 +1,5 @@
 import 'package:altme/app/app.dart';
 import 'package:altme/dashboard/dashboard.dart';
-import 'package:altme/dashboard/home/tab_bar/tokens/models/token_model.dart';
 import 'package:altme/l10n/l10n.dart';
 import 'package:altme/wallet/cubit/wallet_cubit.dart';
 import 'package:altme/withdrawal_tokens/withdrawal_tokens.dart';
